aled

+++ Parcelle.hpp
@@ -13,7 +13,7 @@ private:
     int numero;
     string proprietaire;
     float surface;
-    Polygone<float> forme;
+    Polygone<int> forme;
     int pConstructible;
 
 public:
@@ -25,12 +25,13 @@ public:
     int getNumero(void) const;
     string getProprietaire(void) const;
     float getSurface(void) const;
-    Polygone<float> getForme(void) const;
+    Polygone<int> getForme(void) const;
     int getPConstructible(void) const;
 
     /* Setters */
     void setNumero(int n);
     void setProprietaire(string prop);
-    void setForme(Polygone<float> _forme);
+    void setForme(Polygone<int> _forme);
     void setPConstructible(int pConstructible);
 };
+
